Reseteando valores por defecto
