+++ docker/ckan/files/ckan.ini
@@ -160,19 +160,19 @@ ckan.site_custom_css =
 
 # Configuración del footer
 ckan.footer.email = email@gmail.com
-ckan.footer.telefono = 0351-5553755
-ckan.footer.direccion = Av. Juan Filloy, Ciudad Universitaria.
-ckan.footer.pagina_oficial = https://www.unc.edu.ar/contactenos
+ckan.footer.telefono = 0555-9999999
+ckan.footer.direccion = Av. Nombre Calle, Ciudad.
+ckan.footer.pagina_oficial = https://www.institución.edu.ar/contactenos
 ckan.footer.idioma = español
 ckan.footer.objetivo = Nuestro objetivo es.
 ckan.footer.logo = /logo_footer_unc.png
 
 # Redes sociales
-ckan.footer.facebook = https://www.facebook.com/universidad.nacional.de.cordoba.oficial/
-ckan.footer.twitter = https://twitter.com/unc_cordoba
-ckan.footer.instagram = https://www.instagram.com/uncoficial/
-ckan.footer.youtube = https://www.youtube.com/channel/UCWQHNoiTQHSwbx9nWu5u8-A
-ckan.footer.linkedin = https://www.linkedin.com/groups/1799627/
+ckan.footer.facebook = https://www.facebook.com/
+ckan.footer.twitter = https://twitter.com/
+ckan.footer.instagram = https://www.instagram.com/
+ckan.footer.youtube = https://www.youtube.com/
+ckan.footer.linkedin = https://www.linkedin.com/
 
 ## Resource Views Settings #####################################################
 ckan.views.default_views = datatables_view image_view pdf_view
